requirements.txt
termcolor==1.1.0
cryptography==2.7
numpy==1.16.4
sanic==19.9.0
python-socketio==4.4.0
aiohttp==3.6.2
Pillow==6.2.1
pycryptodome==3.9.4
python-pkcs11==0.5.0
pyOpenSSL==19.1.0
